+++ launch.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+
+# Launcher for etracer, setting library path to point to LSL
+LD_LIBRARY_PATH=./libs ./etracer