+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 * setting speed and mass through course config
 * detect herring collection
 * detecting left/right herrings
+* sending stims through LSL